refactor: 부정확한 컨트롤러 이름 변경
